Refactor application architecture and remove obsolete code
