+++ VoiceInk/App/MenuBar/MenuBarView.swift
@@ -124,8 +124,8 @@ struct MenuBarView: View {
             }
             .keyboardShortcut("c", modifiers: [.command, .shift])
 
-            Button("History") {
-                menuBarManager.openHistoryQuickAccess()
+            Button("Quick History") {
+                menuBarManager.openQuickHistory()
             }
             .keyboardShortcut("h", modifiers: [.command, .shift])
