+++ VoiceInk/Features/Shortcuts/Models/ShortcutAction.swift
@@ -7,7 +7,7 @@ enum ShortcutAction: Hashable {
     case pasteLastEnhancement
     case retryLastTranscription
     case cancelRecorder
-    case openHistoryWindow
+    case openQuickHistory
     case quickAddToDictionary
     case mode(UUID)
     case recorderPanelEscape
@@ -40,7 +40,7 @@ enum ShortcutAction: Hashable {
             return "retryLastTranscription"
         case .cancelRecorder:
             return "cancelRecorder"
-        case .openHistoryWindow:
+        case .openQuickHistory:
             return "openHistoryWindow"
         case .quickAddToDictionary:
             return "quickAddToDictionary"
@@ -67,8 +67,8 @@ enum ShortcutAction: Hashable {
             return String(localized: "Retry Last Transcription")
         case .cancelRecorder:
             return String(localized: "Cancel Recording")
-        case .openHistoryWindow:
-            return String(localized: "Open Quick Access")
+        case .openQuickHistory:
+            return String(localized: "Open Quick History")
         case .quickAddToDictionary:
             return String(localized: "Quick Add to Dictionary")
         case .mode(let id):
@@ -92,7 +92,7 @@ enum ShortcutAction: Hashable {
         .pasteLastTranscription,
         .pasteLastEnhancement,
         .retryLastTranscription,
-        .openHistoryWindow,
+        .openQuickHistory,
         .quickAddToDictionary,
     ]
 
@@ -107,7 +107,7 @@ enum ShortcutAction: Hashable {
         .pasteLastEnhancement,
         .retryLastTranscription,
         .cancelRecorder,
-        .openHistoryWindow,
+        .openQuickHistory,
         .quickAddToDictionary,
     ]
